+++ xformers/profiler/__init__.py
@@ -5,7 +5,6 @@
 
 from .api import profile, step
 from .profiler import MemSnapshotsProfiler, NsightProfiler, PyTorchProfiler
-from .slow_ops_profiler import DetectSlowOpsProfiler
 
 __all__ = [
     "profile",
@@ -13,5 +12,4 @@ __all__ = [
     "MemSnapshotsProfiler",
     "PyTorchProfiler",
     "NsightProfiler",
-    "DetectSlowOpsProfiler",
 ]